Add dashboard node to visualize state machine graph
This implements the long wished-for feature #4 by adding a dashboard
node which shows the state-machine visualization.

The UI node must be connected to the new output node of the smxstate node
through which the full SVG as well as state transitions are transmitted.
The same code as in the editors UI is used to render the SVG. It was just
extended to use a shadow root so the dashboard's CSS doesn't interfere
unexpectedly with the SVG. Instead the theming values are deliberately
carried over and applied to match the visual style.

The code to render the graph was extracted from the HTTP request handler
to the new `renderGraph` function which is now always called when the
node starts. That way an always up-to-date SVG is created. When state
transitions were previously only communicated to the editor they are now
also send through the newly introduced output.