Visualize state machine graph using ui-svg
This implements the long wished-for feature #4 by adding a dashboard
node which shows the state-machine visualization.

A new output port on the state machine node can be connected to a
[node-red-contrib-ui-svg] node. This reuses the already existing node
for any SVG rendering on the dashboard. All necessary messages are sent
to render the SVG in the same way as is already done in the editor's
sidebar.

Despite the attempt to make it work with as little manual intervention
as possible there are a few changes the user must apply on the 'SVG
graphics' node:

1. Enable *Send output msg when the client is (re)reloaded*
2. Change the CSS to:
```
div.ui-svg path {
    fill: none;
}

div.ui-svg text {
    fill: var(--nr-dashboard-widgetTextColor);
}

div.ui-svg svg>g>polygon {
    display: none;
}
```
3. Ideally empty out the prefilled SVG to not see it flashing until the
   state machine's SVG is ready.

[node-red-contrib-ui-svg] https://flows.nodered.org/node/node-red-contrib-ui-svg