Fix race condition in graph rendering
The `renderGraph` function runs asynchronously in the background to
produce the SVG for the state machine. In the mean time the state
machine already runs and performs transitions. Those updates to the SVG,
which show the current state, are published before the actual graph is
ready. Once the graph is ready it is published without the highlights
for the current state.

To the user this looked like the currently active state isn't actually
shown initially.

Fixed by also publishing a `set_style` for the current state after the
full graph was rendered.

This code was just copied from `transitionFcn` but `getStatepaths` was
extracted to not duplicate it.

Also introduced the `machineId` variable as this is what was used in the
loop and the code already made multiple accesses which could just use
the variable instead.