Only run state machine upon explicit request
Until now the state machine logic was started immediately. So for the
whole lifetime of the flow it was running. This caused problems with its
use in multiple independent flows where each can be enabled or not based
on its own decision. With the current implementation it meant a running
state machine continued to run (e.g. update a timer) even though the
flow was supposed to be disabled.

The solution here is to change how the node works. On startup it only
creates the machine without actually running any actions by setting the
`execute` option to `false`. It then requests the graph to be rendered
(which requires a started machine) and immediately stops the machine
again.

Only upon request (via a message having the topic `sm-start`) will the
state machine actually start and perform actions. The analogous
`sm-stop` allows to stop the state machine again. This corresponds to
how the machine is meant to be used in the different flows.