Limit parallel state chart rendering
Since state charts are now an output of the node they are being
generated when a flow is being deployed. When there are many nodes there
will be many parallel calls to `smcat` which can quickly overload a
server which only has a few CPU cores.

Limit the amount of parallel `smcat` instances to not overload the
server. Getting some graphs with a bit of delay is better than a
completely overloaded server.