Fix update-context also changing history's context
The context object was modified directly. Since the same object is also
in the history state modifying the current context object also changed
the context of the history state. To the context listener this looks
like the context didn't change, so not further processing took place.

Fixed by using `Object.assign({}, ...)` which is also what is used
in xstate itself (see `updateContext` in `node_modules/xstate/es/
utils.js`).